Doing some testing, fixing up merged changes
