+++ src/navigation/map_creation/config/cone_placement.yaml
@@ -3,5 +3,5 @@ cone_placement_node:
     view_x: 25.0
     view_y: 25.0
     radius: 1.9
-    min_detections: 2
+    min_detections: 7
     use_sim_time: False
